requirements.txt
pymongo>=4.0.0
bson>=0.5.0
psutil>=5.9.0
matplotlib>=3.7.0
numpy>=1.24.0
pillow>=9.0.0
